feat: adding grpcio-tools to installation process
stubs should be generated when running `pip install -e .`
