+++ setup.cfg
@@ -28,7 +28,7 @@ project_urls =
     Bug Reports = https://github.com/bentoml/BentoML/issues
     BentoML Community Slack = https://bit.ly/2N5IpbB
     BentoML Official Blog = https://modelserving.com
-    BentoML Twitter = http://twitter.com/bentomlai
+    BentoML Twitter = https://twitter.com/bentomlai
 
 [options]
 packages = bentoml
@@ -78,6 +78,7 @@ install_requires =
 python_requires = >=3.7
 include_package_data = True
 setup_requires =
+    grpcio-tools>=1.43.0  # matching with grpcio dependency for generating stubs
     setuptools>=59.0
     setuptools-scm>=6.2.3
     wheel
